cyber-audit-flask/requirements.txt
Flask
Werkzeug
google-cloud-aiplatform
python-dotenv
gunicorn